refactor common logic JenisRealisasi
