+++ src/main/java/cc/kertaskerja/realisasi/domain/JenisRealisasi.java
@@ -1,0 +1,6 @@
+package cc.kertaskerja.realisasi.domain;
+
+public enum JenisRealisasi {
+    NAIK,
+    TURUN
+}
