--- src/main/java/cc/kertaskerja/realisasi_pemda_service/realisasi/domain/JenisRealisasi.java
@@ -1,6 +1,0 @@
-package cc.kertaskerja.realisasi_pemda_service.realisasi.domain;
-
-public enum JenisRealisasi {
-    NAIK,
-    TURUN
-}
