+++ src/main/java/cc/kertaskerja/realisasi_pemda_service/sasaran/domain/SasaranService.java
@@ -1,6 +1,6 @@
 package cc.kertaskerja.realisasi_pemda_service.sasaran.domain;
 
-import cc.kertaskerja.realisasi_pemda_service.realisasi.domain.JenisRealisasi;
+import cc.kertaskerja.realisasi.domain.JenisRealisasi;
 import cc.kertaskerja.realisasi_pemda_service.sasaran.web.SasaranRequest;
 import jakarta.validation.Valid;
 import org.springframework.stereotype.Service;
