+++ src/test/java/cc/kertaskerja/realisasi_pemda_service/tujuan/domain/TujuanRepositoryR2dbcTests.java
@@ -1,7 +1,7 @@
 package cc.kertaskerja.realisasi_pemda_service.tujuan.domain;
 
 import cc.kertaskerja.config.DataConfig;
-import cc.kertaskerja.realisasi_pemda_service.realisasi.domain.JenisRealisasi;
+import cc.kertaskerja.realisasi.domain.JenisRealisasi;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.data.r2dbc.DataR2dbcTest;
