+++ src/test/java/cc/kertaskerja/realisasi_pemda_service/tujuan/domain/TujuanServiceTests.java
@@ -1,6 +1,6 @@
 package cc.kertaskerja.realisasi_pemda_service.tujuan.domain;
 
-import cc.kertaskerja.realisasi_pemda_service.realisasi.domain.JenisRealisasi;
+import cc.kertaskerja.realisasi.domain.JenisRealisasi;
 import cc.kertaskerja.realisasi_pemda_service.tujuan.web.TujuanRequest;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
